feat(core): use startup level to control neovim behavior
I have been looking for a flexible way to control whether Neovim loads
certain plugins or enables some features. Before this commit, I used an
environment variable to decide whether to load the development plugins.

But it isn't flexible enough. Now I have chosen a new approach to
implement it. The 'env' command have a option to change the zeroth
argument. So we can define these aliases:
  alias vi='env -a vi nvim'
  alias vim='env -a vim nvim'
The neovim decide whether to load/enable the plugins/features should
based on this zeroth argument.